fix-heart-emoji
The issue was with the way browsers threat unicode emojis. There are a lot of heart versions. The one that was used in that context was an old one. So it was not parsed as an emoji.

There are two solutions: change the emoji at the client's brand. Parse the heart to the newer one so it's processed as an emoji.